finalize project

+++ customer/consumer.py
@@ -22,27 +22,33 @@ def create_offer(offer):
     offer_model = OffersSerializer(data = offer)
     if offer_model.is_valid():
         offer_model.save()
-        logger.info('offer created successfully')
+        print('offer created successfully')
         return 
-    logger.error('offer not created')
+    print('offer not created')
 
 # The request is to update an offer.
 def update_offer(offer):
+    if 'id' not in offer:
+        print("you forget id in offer")
+        return
     try:
         offer_model = OffersSerializer(data = offer,instance = Offer.objects.get(id = offer['id']))    
         if offer_model.is_valid():
             offer_model.save()
-            logger.info('offer updated successfully')
+            print('offer updated successfully')
     except Offer.DoesNotExist:
-        logger.error('offer Does Not Exist')
+        print('offer Does Not Exist')
 
 # The request is to delete an offer.
 def delete_offer(offer):
+    if 'id' not in offer:
+        print("you forget id in offer")
+        return
     try:
         Offer.objects.get(id = offer['id']).delete()
-        logger.info('offer deleted successfully')
+        print('offer deleted successfully')
     except Offer.DoesNotExist:
-        logger.error('offer Does Not Exist')
+        print('offer Does Not Exist')
 
 def consumer():
     # Create a Kafka consumer.
@@ -61,10 +67,12 @@ def consumer():
     # Loop over the messages.
     for message in consumer:
         # Display the offer on the console.
-        print(message.value)
         offer = message.value
+        print(offer)
         key = offer['type']
         offer.pop('type' , None)
+        print(json.dumps(offer, indent=4, separators=(". ", " = ")))
+        print(key)
         # Determine the request type.
         if key == 'create-offer':
             create_offer(offer)
